FIX-#3605: Address comments related to new project structure
