+++ modin/core/storage_formats/pandas/query_compiler.py
@@ -218,7 +218,7 @@ class PandasQueryCompiler(BaseQueryCompiler):
     Query compiler for the pandas storage format.
 
     This class translates common query compiler API into the DataFrame Algebra
-    queries, that is supposed to be executed by :py:class:`~modin.engines.base.frame.data.PandasDataframe`.
+    queries, that is supposed to be executed by :py:class:`~modin.core.dataframe.pandas.dataframe.dataframe.PandasDataframe`.
 
     Parameters
     ----------
