+++ modin/experimental/cloud/rayscale.py
@@ -69,7 +69,7 @@ class _Immediate:
 
 class RayCluster(BaseCluster):
     target_engine = "Cloudray"
-    target_partition = "Pandas"
+    target_storage_format = "Pandas"
 
     __base_config = os.path.join(
         os.path.abspath(os.path.dirname(__file__)), "ray-autoscaler.yml"
