+++ modin/experimental/core/execution/native/implementations/omnisci_on_native/test/test_dataframe.py
@@ -723,7 +723,7 @@ class TestGroupby:
         run_and_compare(groupby_count, data=self.data, cols=cols, as_index=as_index)
 
     @pytest.mark.xfail(
-        reason="Currently mean() passes a lambda into query compiler which cannot be executed on OmniSci storage format"
+        reason="Currently mean() passes a lambda into query compiler which cannot be executed on OmniSci engine"
     )
     @pytest.mark.parametrize("cols", cols_value)
     @pytest.mark.parametrize("as_index", bool_arg_values)
@@ -1955,7 +1955,7 @@ class TestDropna:
     @pytest.mark.parametrize("dropna", [True, False])
     def test_dropna_groupby(self, by, dropna):
         def applier(df, *args, **kwargs):
-            # OmniSci storage format preserves NaNs at the result of groupby,
+            # OmniSci engine preserves NaNs at the result of groupby,
             # so replacing NaNs with '0' to match with Pandas.
             # https://github.com/modin-project/modin/issues/2878
             return df.groupby(by=by, dropna=dropna).sum().fillna(0)
